requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
redis==5.2.1
aiohttp>=3.9.0
moviepy>=1.0.3
google-auth
google-auth-oauthlib
google-api-python-client
python-jose